refactor: change student folders
